requirements.txt
anyio==4.9.0
APScheduler==3.10.4
cachetools==4.2.2
certifi==2025.7.14
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
